+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.0.0](https://github.com/accessible-ui/disclosure/compare/v2.0.2...v3.0.0) (2020-07-12)
+
+### ⚠ BREAKING CHANGES
+
+- rename useTarget() to useA11yTarget(), useTrigger() to useA11yTrigger(),
+  useCloseButton() to useA11yCloseButton()
+
+- rename use**_ hooks to useA11y_** ([0a0eb67](https://github.com/accessible-ui/disclosure/commit/0a0eb6743bdf658ea2b1fd4ddcd20fe3831e1bc5))
+
 ### [2.0.2](https://github.com/accessible-ui/disclosure/compare/v2.0.1...v2.0.2) (2020-07-12)
 
 ### Bug Fixes